# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GreenLink & Sunrise Agrovet Retail Operations OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: PIN login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: No session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/verify-pin:
    post:
      operationId: verifyOwnerPin
      tags: [auth]
      summary: Verify owner PIN for privileged action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPinInput"
      responses:
        "200":
          description: PIN verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── SHOPS ───────────────────────────────────────────────
...
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: List debts
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status (unpaid|partial|paid)
        - name: q
          in: query
          schema:
            type: string
          description: Search by customer name/phone
      responses:
        "200":
          description: Debts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
    post:
      operationId: createDebt
      tags: [debts]
      summary: Create debt record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtInput"
      responses:
        "201":
          description: Debt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{debtId}:
    get:
      operationId: getDebt
      tags: [debts]
      summary: Get debt with payment history
      parameters:
        - name: debtId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtWithPayments"
    patch:
      operationId: updateDebt
      tags: [debts]
      summary: Update debt
      parameters:
        - name: debtId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtUpdate"
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{debtId}/payments:
    post:
      operationId: recordDebtPayment
      tags: [debts]
      summary: Record a payment against a debt
      parameters:
        - name: debtId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtPaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPayment"

  # ─── INVENTORY MOVEMENTS ─────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: unreadOnly
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── AUDIT LOG ───────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ─── AUTH ─────────────────────────────────────────────
    AuthToken:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
        shopId:
          type: string
        shopName:
          type: string
        userName:
          type: ["string", "null"]
      required: [token, role, shopId, shopName]
    LoginInput:
      type: object
      properties:
        shopId:
          type: string
        role:
          type: string
          enum: [owner, cashier]
        pin:
          type: string
        userName:
          type: ["string", "null"]
      required: [shopId, role, pin]
    Session:
      type: object
      properties:
        role:
          type: string
        shopId:
          type: string
        shopName:
          type: string
        userName:
          type: ["string", "null"]
      required: [role, shopId, shopName]
    VerifyPinInput:
      type: object
      properties:
        shopId:
          type: string
        pin:
          type: string
      required: [shopId, pin]

    # ─── SHOPS ────────────────────────────────────────────
...
    Debt:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        saleId:
          type: ["string", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        paidAt:
          type: ["string", "null"]
      required: [id, shopId, customerName, customerPhone, totalAmount, amountPaid, balance, status, createdAt]
    DebtPayment:
      type: object
      properties:
        id:
          type: string
        debtId:
          type: string
        amount:
          type: number
        paidAt:
          type: string
        recordedBy:
          type: ["string", "null"]
      required: [id, debtId, amount, paidAt]
    DebtWithPayments:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        saleId:
          type: ["string", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        paidAt:
          type: ["string", "null"]
        payments:
          type: array
          items:
            $ref: "#/components/schemas/DebtPayment"
      required: [id, shopId, customerName, customerPhone, totalAmount, amountPaid, balance, status, createdAt, payments]
    DebtInput:
      type: object
      properties:
        shopId:
          type: string
        saleId:
          type: ["string", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
      required: [shopId, customerName, customerPhone, totalAmount]
    DebtUpdate:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
    DebtPaymentInput:
      type: object
      properties:
        amount:
          type: number
        recordedBy:
          type: ["string", "null"]
      required: [amount]

    # ─── INVENTORY ────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        productId:
          type: ["string", "null"]
        debtId:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, shopId, type, title, message, isRead, createdAt]

    # ─── AUDIT ────────────────────────────────────────────
...